# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Reconciler API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface ManualSaleInput {
  /** ISO date string YYYY-MM-DD */
  saleDate: string;
  item: string;
  qty: number;
  rate: number;
  amount: number;
}

export interface ManualPurchaseInput {
  billDate: string;
  purchaseDate: string;
  item: string;
  qty: number;
  rate: number;
  amount: number;
}

export type DeleteByDateRequestType =
  (typeof DeleteByDateRequestType)[keyof typeof DeleteByDateRequestType];

export const DeleteByDateRequestType = {
  sale: "sale",
  purchase: "purchase",
} as const;

export interface DeleteByDateRequest {
  /** ISO date YYYY-MM-DD */
  date: string;
  type: DeleteByDateRequestType;
}

export type SaleRowStatus = (typeof SaleRowStatus)[keyof typeof SaleRowStatus];

export const SaleRowStatus = {
  Matched: "Matched",
  Pending: "Pending",
} as const;

export interface SaleRow {
  id?: number;
  saleDate: string;
  item: string;
  qty: number;
  rate: number;
  amount: number;
  purchaseBillDate?: string | null;
  status: SaleRowStatus;
}

export type PurchaseRowStatus =
  (typeof PurchaseRowStatus)[keyof typeof PurchaseRowStatus];

export const PurchaseRowStatus = {
  Matched: "Matched",
  Unmatched: "Unmatched",
  Extra: "Extra",
} as const;

export interface PurchaseRow {
  id?: number;
  billDate: string;
  purchaseDate: string;
  item: string;
  qty: number;
  rate: number;
  amount: number;
  status: PurchaseRowStatus;
}

export interface ItemSummary {
  item: string;
  salesQty: number;
  salesAmount: number;
  purchaseQty: number;
  purchaseAmount: number;
  pendingQty: number;
  pendingAmount: number;
}

export interface ReconciliationResult {
  salesRows: SaleRow[];
  purchaseRows: PurchaseRow[];
  summary: ItemSummary[];
  matchedCount: number;
  pendingCount: number;
  unmatchedPurchaseCount: number;
}

export type RunReconciliationBody = {
  salesFile?: Blob;
  purchaseFile?: Blob;
};