# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stock Reconciler API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: reconciliation
    description: Reconciliation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current auth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"

  /reconciliation/run:
    post:
      operationId: runReconciliation
      tags: [reconciliation]
      summary: Upload file(s) and run strict lot matching
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                salesFile:
                  type: string
                  format: binary
                purchaseFile:
                  type: string
                  format: binary
      responses:
        "200":
          description: Reconciliation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reconciliation/reports:
    get:
      operationId: getReports
      tags: [reconciliation]
      summary: Get all saved records for current user
      responses:
        "200":
          description: Full reconciliation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"

  /reconciliation/records/sale:
    post:
      operationId: addSaleRecord
      tags: [reconciliation]
      summary: Add a single sale record manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualSaleInput"
      responses:
        "200":
          description: Record added, returns updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reconciliation/records/purchase:
    post:
      operationId: addPurchaseRecord
      tags: [reconciliation]
      summary: Add a single purchase record manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualPurchaseInput"
      responses:
        "200":
          description: Record added, returns updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reconciliation/records/sale/{id}:
    delete:
      operationId: deleteSaleRecord
      tags: [reconciliation]
      summary: Delete a specific sale record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Record deleted, returns updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reconciliation/records/purchase/{id}:
    delete:
      operationId: deletePurchaseRecord
      tags: [reconciliation]
      summary: Delete a specific purchase record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Record deleted, returns updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reconciliation/records/date:
    delete:
      operationId: deleteByDate
      tags: [reconciliation]
      summary: Delete all records for a specific date (sales or purchases)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteByDateRequest"
      responses:
        "200":
          description: Records deleted, returns updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"

  /reconciliation/download/{fileType}:
    post:
      operationId: downloadFile
      tags: [reconciliation]
      summary: Download a reconciliation report as Excel
      parameters:
        - name: fileType
          in: path
          required: true
          schema:
            type: string
            enum:
              - updated-sales
              - pending-pavati
              - datewise-report
              - purchase-exceptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconciliationResult"
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required: [id]

    GetCurrentAuthUserResponse:
      type: object
      properties:
        user:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required: [user]

    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
          nullable: true
      required: [code, code_verifier, redirect_uri, state]

    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required: [token]

    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    ManualSaleInput:
      type: object
      properties:
        saleDate:
          type: string
          description: "ISO date string YYYY-MM-DD"
        item:
          type: string
        qty:
          type: number
        rate:
          type: number
        amount:
          type: number
      required: [saleDate, item, qty, rate, amount]

    ManualPurchaseInput:
      type: object
      properties:
        billDate:
          type: string
        purchaseDate:
          type: string
        item:
          type: string
        qty:
          type: number
        rate:
          type: number
        amount:
          type: number
      required: [billDate, purchaseDate, item, qty, rate, amount]

    DeleteByDateRequest:
      type: object
      properties:
        date:
          type: string
          description: "ISO date YYYY-MM-DD"
        type:
          type: string
          enum: [sale, purchase]
      required: [date, type]

    SaleRow:
      type: object
      properties:
        id:
          type: integer
        saleDate:
          type: string
        item:
          type: string
        qty:
          type: number
        rate:
          type: number
        amount:
          type: number
        purchaseBillDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [Matched, Pending]
      required: [saleDate, item, qty, rate, amount, status]

    PurchaseRow:
      type: object
      properties:
        id:
          type: integer
        billDate:
          type: string
        purchaseDate:
          type: string
        item:
          type: string
        qty:
          type: number
        rate:
          type: number
        amount:
          type: number
        status:
          type: string
          enum: [Matched, Unmatched, Extra]
      required: [billDate, purchaseDate, item, qty, rate, amount, status]

    ItemSummary:
      type: object
      properties:
        item:
          type: string
        salesQty:
          type: number
        salesAmount:
          type: number
        purchaseQty:
          type: number
        purchaseAmount:
          type: number
        pendingQty:
          type: number
        pendingAmount:
          type: number
      required: [item, salesQty, salesAmount, purchaseQty, purchaseAmount, pendingQty, pendingAmount]

    ReconciliationResult:
      type: object
      properties:
        salesRows:
          type: array
          items:
            $ref: "#/components/schemas/SaleRow"
        purchaseRows:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseRow"
        summary:
          type: array
          items:
            $ref: "#/components/schemas/ItemSummary"
        matchedCount:
          type: integer
        pendingCount:
          type: integer
        unmatchedPurchaseCount:
          type: integer
      required: [salesRows, purchaseRows, summary, matchedCount, pendingCount, unmatchedPurchaseCount]