# lib/api-zod/src/generated/types/deleteByDateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Reconciler API
 * OpenAPI spec version: 0.1.0
 */
import type { DeleteByDateRequestType } from "./deleteByDateRequestType";

export interface DeleteByDateRequest {
  /** ISO date YYYY-MM-DD */
  date: string;
  type: DeleteByDateRequestType;
}