# lib/api-zod/src/generated/types/itemSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Reconciler API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemSummary {
  item: string;
  salesQty: number;
  salesAmount: number;
  purchaseQty: number;
  purchaseAmount: number;
  pendingQty: number;
  pendingAmount: number;
}