# lib/api-zod/src/generated/types/manualSaleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Reconciler API
 * OpenAPI spec version: 0.1.0
 */

export interface ManualSaleInput {
  /** ISO date string YYYY-MM-DD */
  saleDate: string;
  item: string;
  qty: number;
  rate: number;
  amount: number;
}