# lib/api-zod/src/generated/types/reconciliationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Reconciler API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemSummary } from "./itemSummary";
import type { PurchaseRow } from "./purchaseRow";
import type { SaleRow } from "./saleRow";

export interface ReconciliationResult {
  salesRows: SaleRow[];
  purchaseRows: PurchaseRow[];
  summary: ItemSummary[];
  matchedCount: number;
  pendingCount: number;
  unmatchedPurchaseCount: number;
}