# lib/api-zod/src/generated/types/saleRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Reconciler API
 * OpenAPI spec version: 0.1.0
 */
import type { SaleRowStatus } from "./saleRowStatus";

export interface SaleRow {
  id?: number;
  saleDate: string;
  item: string;
  qty: number;
  rate: number;
  amount: number;
  purchaseBillDate?: string | null;
  status: SaleRowStatus;
}